refactor(scraping): extract scrape models and parsers into dedicated modules
